pyproject.toml
[project]
name = "titanium"
version = "2.0.0"
description = "Multipurpose Discord Bot"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.30.0",
    "atlas-provider-sqlalchemy>=0.4.0",
    "colorthief>=0.2.1",
    "discord-py>=2.6.4",
    "dotenv>=0.9.9",
    "emoji>=2.15.0",
    "humanize>=4.13.0",
    "jinja2>=3.1.6",
    "pillow>=11.3.0",
    "pydantic>=2.12.4",
    "pytest-playwright>=0.7.1",
    "qrcode[pil]>=8.2",
    "shortuuid>=1.0.13",
    "spotipy>=2.25.1",
    "sqlalchemy[asyncio]>=2.0.41",
    "wand>=0.6.13",
]

[dependency-groups]
dev = [
    "ruff>=0.12.5",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = ["I"]

[tool.basedpyright]
typeCheckingMode = "basic"
